app폴더 내의 page.tsx에 적은 html과 css를 components폴더로 이동
